requirements.txt
Django==1.9.1
coverage==3.7.1
dj-database-url==0.3.0
django-allauth==0.24.1
django-braces==1.4.0
django-crispy-forms==1.6.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.4
django-environ==0.3.0
django-extensions==1.6.1
django-floppyforms==1.6.0
django-imagekit==3.3
django-paypal==0.3.2
django-suit==0.2.15
dj-static==0.0.6
psycopg2==2.5.4
python-dateutil==2.4.0
model-mommy==1.2.6
mock==1.0.1
django-ckeditor==5.0.2
pillow==2.7.0
gunicorn==19.3.0
static3==0.5.1
